+++ CHANGELOG.md
@@ -1,6 +1,43 @@
 # Changelog
-## 1.40.5 (2025-01-22)
+## 1.41.2 (2025-01-22)
 - Add Jaarrekening PEVA form [DL-6284]
+## 1.41.1 (2025-01-22)
+### General
+- Fix configuration in subjectsAndPaths.js
+
+### Deploy notes
+#### docker commands
+```
+drc restart vendor-data-distribution
+```
+
+## 1.41.0 (2025-01-21)
+
+### General
+- Add submissions cross referencing components and service
+
+### Deploy notes
+#### docker-compose.override.yml
+##### worship-decisions-cross-reference
+
+Ensure the environment variables are correctly set for `worship-decisions-cross-reference`, e.g. :
+
+```
+worship-decisions-cross-reference:
+  environment:
+    WORSHIP_DECISIONS_BASE_URL: "https://databankerediensten.lokaalbestuur.vlaanderen.be/search/submissions/"
+```
+The following links;
+- DEV: "https://dev.databankerediensten.lokaalbestuur.lblod.info/search/submissions/"
+- QA: "https://databankerediensten.lokaalbestuur.lblod.info/search/submissions/"
+- PROD: "https://databankerediensten.lokaalbestuur.vlaanderen.be/search/submissions/"
+
+docker commands
+
+```
+drc restart migrations
+drc up -d worship-decisions-cross-reference frontend
+```
 ## 1.40.4 (2024-12-13)
 - New semantic form `Kerkenbeleidsplan`
 - New semantic forms for cross referencing